When we launch instruments w/o deleting the content and settings folder in xcode 6, it fails to launch the app with the error
Instruments Trace Error : Target failed to run: The operation couldn’t be completed. (FBSOpenApplicationErrorDomain error 8.) : Failed to launch process with bundle identifier '<bundle name>'

As a fix we launch instruments twice.

As an alternate we can try simctl to install app, launch app and then run instruments.
But launching instruments twice seems to be an easier workaround and let's use it
until Apple fixes it.