Remove `scrollToVisible()` call in `isStale()`.
While `scrollToVisible()` provides signal when the element is stale, when it's not (i.e. it is valid and on screen), it can move the element or tap it in some cases.
I've filed an Apple bug (rdar://18735706) for this behaviour but in the meantime, it seems like we should remove our usage of `scrollToVisible()` from `isStale()`.